feat: don't timeout icon load on client, only on ssr
The timeout was introduced mainly to not have the SSR hanging forever. At the client side, sometime due to the network conditions, the timeout become a footgun that failed to render the icon even the data is eventually received. This PR changes that to have `fetchTimeout` only effective at SSR. This also align more between the CSS and SSR modes.