Update README.md to include a Nuxt 4 app dir note
I noticed that there wasn't any documentation on configuring Nuxt Icons with the new Nuxt 4 app directory. It took me a while to debug until I realized the issue.

This could be a temporary fix, but the nuxt-icon module can detect the folder structure to determine where the assets directory really is.